+++ src/natten/backends/__init__.py
@@ -31,9 +31,8 @@ logger = log.get_logger(__name__)
 import torch  # noqa: F401
 from torch import Tensor
 
-from .blackwell_fmha import can_run_cutlass_blackwell_fmha, cutlass_blackwell_fmha
+from .blackwell_fmha import cutlass_blackwell_fmha
 from .blackwell_fna import (
-    can_run_cutlass_blackwell_fna,
     cutlass_blackwell_fna_generic,
     na1d_cutlass_blackwell_fna,
     na2d_cutlass_blackwell_fna,
@@ -53,25 +52,25 @@ from .configs import (
     get_configs_for_flex_fna,
 )
 
-from .flex import (
+from .configs.checks import (
+    can_run_cutlass_blackwell_fmha,
+    can_run_cutlass_blackwell_fna,
+    can_run_cutlass_fna,
+    can_run_cutlass_hopper_fmha,
+    can_run_cutlass_hopper_fna,
     can_run_flex_attention,
-    flex_fmha,
-    flex_fna_generic,
-    na1d_flex,
-    na2d_flex,
-    na3d_flex,
 )
+
+from .flex import flex_fmha, flex_fna_generic, na1d_flex, na2d_flex, na3d_flex
 from .fmha import can_run_cutlass_fmha, cutlass_fmha
 from .fna import (
-    can_run_cutlass_fna,
     cutlass_fna_generic,
     na1d_cutlass_fna,
     na2d_cutlass_fna,
     na3d_cutlass_fna,
 )
-from .hopper_fmha import can_run_cutlass_hopper_fmha, cutlass_hopper_fmha
+from .hopper_fmha import cutlass_hopper_fmha
 from .hopper_fna import (
-    can_run_cutlass_hopper_fna,
     cutlass_hopper_fna_generic,
     na1d_cutlass_hopper_fna,
     na2d_cutlass_hopper_fna,
@@ -79,20 +78,22 @@ from .hopper_fna import (
 )
 
 
-def choose_backend(input_tensor: Tensor, torch_compile: bool) -> str:
-    if can_run_cutlass_blackwell_fna(input_tensor):
+def choose_backend(
+    query: Tensor, key: Tensor, value: Tensor, torch_compile: bool
+) -> str:
+    if can_run_cutlass_blackwell_fna(query, key, value):
         logger.debug("Backend not set; picked Blackwell FNA kernel.")
         return "blackwell-fna"
 
-    if can_run_cutlass_hopper_fna(input_tensor):
+    if can_run_cutlass_hopper_fna(query, key, value):
         logger.debug("Backend not set; picked Hopper FNA kernel.")
         return "hopper-fna"
 
-    if can_run_cutlass_fna(input_tensor):
+    if can_run_cutlass_fna(query, key, value):
         logger.debug("Backend not set; picked CUTLASS (2.X) FNA kernel.")
         return "cutlass-fna"
 
-    if can_run_flex_attention(input_tensor, torch_compile=torch_compile):
+    if can_run_flex_attention(query, key, value, torch_compile=torch_compile):
         logger.debug("Backend not set; picked Flex Attention kernel.")
         return "flex-fna"
 
@@ -102,20 +103,22 @@ def choose_backend(input_tensor: Tensor, torch_compile: bool) -> str:
     )
 
 
-def choose_fmha_backend(input_tensor: Tensor, torch_compile: bool) -> str:
-    if can_run_cutlass_blackwell_fmha(input_tensor):
+def choose_fmha_backend(
+    query: Tensor, key: Tensor, value: Tensor, torch_compile: bool
+) -> str:
+    if can_run_cutlass_blackwell_fmha(query, key, value):
         logger.debug("Backend not set; picked Blackwell FMHA kernel.")
         return "blackwell-fmha"
 
-    if can_run_cutlass_hopper_fmha(input_tensor):
+    if can_run_cutlass_hopper_fmha(query, key, value):
         logger.debug("Backend not set; picked Hopper FMHA kernel.")
         return "hopper-fmha"
 
-    if can_run_cutlass_fmha(input_tensor):
+    if can_run_cutlass_fmha(query, key, value):
         logger.debug("Backend not set; picked CUTLASS (2.X) FMHA kernel.")
         return "cutlass-fmha"
 
-    if can_run_flex_attention(input_tensor, torch_compile=torch_compile):
+    if can_run_flex_attention(query, key, value, torch_compile=torch_compile):
         logger.debug("Backend not set; picked Flex Attention kernel.")
         return "flex-fmha"
 
@@ -125,37 +128,39 @@ def choose_fmha_backend(input_tensor: Tensor, torch_compile: bool) -> str:
     )
 
 
-def get_compatible_backends(input_tensor: Tensor, torch_compile: bool) -> List[str]:
+def get_compatible_backends(
+    query: Tensor, key: Tensor, value: Tensor, torch_compile: bool
+) -> List[str]:
     compatible_backends = []
-    if can_run_cutlass_blackwell_fna(input_tensor):
+    if can_run_cutlass_blackwell_fna(query, key, value):
         compatible_backends.append("blackwell-fna")
 
-    if can_run_cutlass_hopper_fna(input_tensor):
+    if can_run_cutlass_hopper_fna(query, key, value):
         compatible_backends.append("hopper-fna")
 
-    if can_run_cutlass_fna(input_tensor):
+    if can_run_cutlass_fna(query, key, value):
         compatible_backends.append("cutlass-fna")
 
-    if can_run_flex_attention(input_tensor, torch_compile=torch_compile):
+    if can_run_flex_attention(query, key, value, torch_compile=torch_compile):
         compatible_backends.append("flex-fna")
 
     return compatible_backends
 
 
 def get_compatible_fmha_backends(
-    input_tensor: Tensor, torch_compile: bool
+    query: Tensor, key: Tensor, value: Tensor, torch_compile: bool
 ) -> List[str]:
     compatible_backends = []
-    if can_run_cutlass_blackwell_fmha(input_tensor):
+    if can_run_cutlass_blackwell_fmha(query, key, value):
         compatible_backends.append("blackwell-fmha")
 
-    if can_run_cutlass_hopper_fmha(input_tensor):
+    if can_run_cutlass_hopper_fmha(query, key, value):
         compatible_backends.append("hopper-fmha")
 
-    if can_run_cutlass_fmha(input_tensor):
+    if can_run_cutlass_fmha(query, key, value):
         compatible_backends.append("cutlass-fmha")
 
-    if can_run_flex_attention(input_tensor, torch_compile=torch_compile):
+    if can_run_flex_attention(query, key, value, torch_compile=torch_compile):
         compatible_backends.append("flex-fmha")
 
     return compatible_backends